complete

+++ FallingSky[Rebuild]/HighScore.h
@@ -13,7 +13,7 @@ public:
 	HighScore();
 	~HighScore();
 	void ReadFile(RenderWindow &window);
-	void WriteFile(string name, float Score);
+	void WriteFile(string name, int Score);
 private:
-	vector<pair<float, string>> scoreboard;
+	vector<pair<int, string>> scoreboard;
 };
